requirements.txt
google-adk==0.1.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0
fastapi>=0.115.0
uvicorn>=0.27.0
python-jose==3.3.0
passlib==1.7.4
passlib[bcrypt]==1.7.4
